pyproject.toml
[project]
name = "aisearch-openai-rag-audio-communication-services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.9.3",
    "azure-identity==1.18.0",
    "azure-search-documents==11.6.0b4",
    "python-dotenv==1.0.1",
    "azure-storage-blob==12.23.1",
    "gunicorn>=23.0.0",
    "rich>=13.9.4",
    "azure-eventgrid>=4.21.0",
    "azure-communication-callautomation==1.4.0b1",
]

[tool.ruff]
target-version = "py39"
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["E501", "E701"]
src = ["app/backend"]
